backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
flask-limiter==4.1.1
pymongo==4.16.0
python-dotenv==1.2.2
PyJWT==2.12.1
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
google-genai==1.10.0
openai==2.30.0
presidio-analyzer==2.2.362
presidio-anonymizer==2.2.362
spacy==3.8.14
sentence-transformers==4.1.0
gunicorn==23.0.0
